+++ gatsby/src/components/buttons.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as styles from "../styles/buttons.module.css"
 import * as visibilityStyles from "../styles/utilities/visibility.module.css"
 import * as orientationStyles from "../styles/utilities/orientation.module.css"
-import { SubNavProps, SubNavBar } from "./nav/subnav.tsx"
+import { SubNavProps, SubNavBar } from "./nav/SubNav.tsx"
 import { StyleProps } from "../utils"
 import { SvgRightArrow } from "./icons.tsx"
 import { HorizontalRule } from "./hr.tsx"
